jasmin-account-api-client/pyproject.toml
[tool.poetry]
name = "jasmin-account-api-client"
version = "1.0.0"
description = "A client library for accessing JASMIN Account API"

authors = ["Alexander Manning <alexander.manning@stfc.ac.uk>"]

readme = "README.md"
packages = [
    {include = "jasmin_account_api_client"},
]
include = ["CHANGELOG.md", "jasmin_account_api_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.22.0"
attrs = "^21.4.0"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 88
profile = "black"